+++ playwright/integrations/authenticated-chat-with-persistent-chat.spec.ts
@@ -170,4 +170,63 @@ test.describe('AuthenticatedChat @AuthenticatedChatWithPersistentChat', () => {
         expect(sessionInitRequestHeaders['authenticatedusertoken']).toBe(authToken);
         expect(sessionInitResponse.status()).toBe(200);
     });
+
+    test("ChatSDK.startChat() should not have any reconnect id if there's no existing chat session", async ({ page }) => {
+        await page.goto(testPage);
+
+        const [reconnectableChatsRequest, reconnectableChatsResponse, runtimeContext] = await Promise.all([
+            page.waitForRequest(request => {
+                return request.url().includes(OmnichannelEndpoints.LiveChatAuthReconnectableChats);
+            }),
+            page.waitForResponse(response => {
+                return response.url().includes(OmnichannelEndpoints.LiveChatAuthReconnectableChats);
+            }),
+            await page.evaluate(async ({ omnichannelConfig, authUrl }) => {
+                const { OmnichannelChatSDK_1: OmnichannelChatSDK, uuidv4 } = window;
+
+                const data = {
+                    contactid: uuidv4() // Ensures it's a new user
+                };
+
+                const payload = {
+                    method: "POST",
+                    body: JSON.stringify(data)
+                };
+
+                const response = await fetch(authUrl, payload);
+                const authToken = await response.text();
+
+                const chatSDKConfig = {
+                    getAuthToken: () => authToken,
+                    persistentChat: {
+                        disable: false
+                    },
+                };
+
+                const chatSDK = new OmnichannelChatSDK.default(omnichannelConfig, chatSDKConfig);
+
+                await chatSDK.initialize();
+
+                const runtimeContext = {
+                    requestId: chatSDK.requestId,
+                    authToken
+                };
+
+                await chatSDK.startChat();
+
+                await chatSDK.endChat();
+
+                return runtimeContext;
+            }, { omnichannelConfig, authUrl })
+        ]);
+
+        const { authToken } = runtimeContext;
+        const reconnectableChatsRequestUrl = `${omnichannelConfig.orgUrl}/${OmnichannelEndpoints.LiveChatAuthReconnectableChats}/${omnichannelConfig.orgId}/${omnichannelConfig.widgetId}/${omnichannelConfig.orgId}?channelId=lcw`;
+
+        const reconnectableChatsRequestHeaders = reconnectableChatsRequest.headers();
+
+        expect(reconnectableChatsRequest.url() === reconnectableChatsRequestUrl).toBe(true);
+        expect(reconnectableChatsRequestHeaders['authenticatedusertoken']).toBe(authToken);
+        expect(reconnectableChatsResponse.status()).toBe(204);
+    });
 });
